chore(all): moved dto and request files to the appropriate folders, change: folder structure of the exception
