+++ backend/src/main/java/project/backend/carts/cartItem/CartItemService.java
@@ -7,7 +7,7 @@ import project.backend.carts.cart.CartService;
 import project.backend.carts.cart.CartStatus;
 import project.backend.courses.course.model.Course;
 import project.backend.courses.course.service.CourseService;
-import project.backend.exception.ResourceNotFoundException;
+import project.backend.exception.types.ResourceNotFoundException;
 import project.backend.user.User;
 import project.backend.user.UserService;
