+++ backend/src/main/java/project/backend/courses/category/service/CategoryServiceImpl.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import project.backend.courses.category.repository.CategoryRepository;
 import project.backend.courses.category.model.Category;
-import project.backend.exception.ResourceNotFoundException;
+import project.backend.exception.types.ResourceNotFoundException;
 
 import java.util.List;
