+++ backend/src/main/java/project/backend/courses/course/dto/FilterCourseDTO.java
@@ -1,4 +1,6 @@
-package project.backend.courses.course.model;
+package project.backend.courses.course.dto;
+
+import project.backend.courses.course.model.Course;
 
 import java.util.List;
