+++ backend/src/main/java/project/backend/courses/language/service/LanguageServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import project.backend.courses.language.repository.LanguageRepository;
 import project.backend.courses.language.model.Language;
-import project.backend.exception.ResourceNotFoundException;
+import project.backend.exception.types.ResourceNotFoundException;
 
 import java.util.List;
