+++ backend/src/main/java/project/backend/courses/lesson/service/CourseLessonService.java
@@ -1,7 +1,7 @@
 package project.backend.courses.lesson.service;
 
 import project.backend.courses.lesson.model.Lesson;
-import project.backend.courses.lesson.model.LessonRequest;
+import project.backend.courses.lesson.request.LessonRequest;
 
 public interface CourseLessonService {
     Lesson addLessonToCourse(Long courseId, LessonRequest lesson);
