+++ backend/src/main/java/project/backend/exception/handler/RestExceptionHandler.java
@@ -1,4 +1,4 @@
-package project.backend.exception;
+package project.backend.exception.handler;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
+import project.backend.exception.response.ExceptionResponse;
+import project.backend.exception.types.ResourceNotFoundException;
 
 import java.sql.Timestamp;
